safari-600.7-branch/LayoutTests/accessibility/tabindex-elements-are-accessible.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../resources/js-test-pre.js"></script>
</head>
<body>

<span tabindex="-1" id="item1">test</span>
<span tabindex="0" id="item2">test</span>
<span tabindex="1" id="item3">test</span>

<p tabindex="-1" id="item4">test


<p id="description"></p>
<div id="console"></div>

<script>

    description("This tests that if tabindex is set on an object, it will be accessible (even if it is usually not)");

    if (window.accessibilityController) {
        for (var k = 1; k < 5; k++) {
            var item = accessibilityController.accessibleElementById("item" + k);
            debug("Item " + k + ": Is focusable: " + item.isFocusable);
            debug("Item " + k + ": " + item.role);
            debug("\n");
        }
    }

</script>

<script src="../resources/js-test-post.js"></script>
</body>
</html>
