safari-600.7-branch/LayoutTests/accessibility/tabindex-elements-are-accessible-expected.txt

test test test
test

This tests that if tabindex is set on an object, it will be accessible (even if it is usually not)

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Item 1: Is focusable: true
Item 1: AXRole: AXGroup


Item 2: Is focusable: true
Item 2: AXRole: AXGroup


Item 3: Is focusable: true
Item 3: AXRole: AXGroup


Item 4: Is focusable: true
Item 4: AXRole: AXGroup


PASS successfullyParsed is true

TEST COMPLETE
